ai-trainer/backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.2.0
python-decouple==3.8
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.0
Pillow==10.0.0
python-docx==0.8.11
PyPDF2==3.0.1
google-generativeai==0.3.0
groq==0.4.1
requests==2.31.0
celery==5.3.1
redis==5.0.0

gunicorn==21.2.0
whitenoise==6.5.0
python-multipart==0.0.6
gTTS==2.3.2
librosa==0.10.0
soundfile==0.12.1
